+++ R/llog.R
@@ -19,13 +19,7 @@
 #'
 #' The functions are wrappers to export the identical functions from the FAdist package.
 #'
-#' @param x,q	vector of quantiles.
-#' @param p	vector of probabilities.
-#' @param n	number of observations.
-#' @param shape	shape parameter.
-#' @param scale scale parameter.
-#' @param log,log.p logical; if TRUE, probabilities p are given as log(p).
-#' @param lower.tail	logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].
+#' @inheritParams params
 #' @return
 #' dllog gives the density, pllog gives the distribution function,
 #' qllog gives the quantile function, and rllog generates random deviates.