+++ R/params.R
@@ -5,6 +5,15 @@
 #' @param ic A string specifying the information-theoretic criterion to use for model averaging ('aic', 'aicc' or 'bic').
 #' @param ci A flag specifying whether to estimate confidence intervals by parametric bootstrapping.
 #' @param ... Unused.
+#' @param x,q	vector of quantiles.
+#' @param p	vector of probabilities.
+#' @param n	number of observations.
+#' @param shape	shape parameter.
+#' @param scale scale parameter.
+#' @param lshape log shape parameter.
+#' @param lscale log scale parameter.
+#' @param log,log.p logical; if TRUE, probabilities p are given as log(p).
+#' @param lower.tail	logical; if TRUE (default), probabilities are P[X <= x],otherwise, P[X > x].
 #' @keywords internal
 #' @name params
 NULL